connects-llm/pyproject.toml
[project]
name = "connects-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ruff>=0.14.5",
]
